+++ skoolkit/sna2ctl.py
@@ -17,6 +17,7 @@
 import argparse
 
 from skoolkit import integer, VERSION
+from skoolkit.config import get_config, show_config, update_options
 from skoolkit.snactl import generate_ctls, write_ctl
 from skoolkit.snapshot import make_snapshot
 
@@ -28,6 +29,7 @@ def run(snafile, options):
     write_ctl(ctls, options.ctl_hex)
 
 def main(args):
+    config = get_config('sna2ctl')
     parser = argparse.ArgumentParser(
         usage='sna2ctl.py [options] FILE',
         description="Generate a control file for a binary (raw memory) file or a SNA, SZX or Z80 snapshot. "
@@ -38,9 +40,11 @@ def main(args):
     group = parser.add_argument_group('Options')
     group.add_argument('-e', '--end', dest='end', metavar='ADDR', type=integer, default=END,
                        help='Stop at this address (default={}).'.format(END))
-    group.add_argument('-h', '--hex', dest='ctl_hex', action='store_const', const=2, default=0,
+    group.add_argument('-h', '--hex', dest='ctl_hex', action='store_const', const=2, default=config['Hex'],
                        help='Write upper case hexadecimal addresses.')
-    group.add_argument('-l', '--hex-lower', dest='ctl_hex', action='store_const', const=1, default=0,
+    group.add_argument('-I', '--ini', dest='params', metavar='p=v', action='append', default=[],
+                       help="Set the value of the configuration parameter 'p' to 'v'. This option may be used multiple times.")
+    group.add_argument('-l', '--hex-lower', dest='ctl_hex', action='store_const', const=1, default=config['Hex'],
                        help='Write lower case hexadecimal addresses.')
     group.add_argument('-m', '--map', dest='code_map', metavar='FILE',
                        help='Use FILE as a code execution map.')
@@ -48,12 +52,17 @@ def main(args):
                        help='Specify the origin address of a binary file (default: 65536 - length).')
     group.add_argument('-p', '--page', dest='page', metavar='PAGE', type=int, choices=list(range(8)),
                        help='Specify the page (0-7) of a 128K snapshot to map to 49152-65535.')
+    group.add_argument('--show-config', dest='show_config', action='store_true',
+                       help="Show configuration parameter values.")
     group.add_argument('-s', '--start', dest='start', metavar='ADDR', type=integer, default=0,
                        help='Start at this address (default=16384).')
     group.add_argument('-V', '--version', action='version', version='SkoolKit {}'.format(VERSION),
                        help='Show SkoolKit version number and exit.')
 
     namespace, unknown_args = parser.parse_known_args(args)
+    if namespace.show_config:
+        show_config('sna2ctl', config)
     if unknown_args or namespace.snafile is None:
         parser.exit(2, parser.format_help())
+    update_options('sna2ctl', namespace, namespace.params, config)
     run(namespace.snafile, namespace)